refactor: change var to constant and add document
